Removed /login route now, will be in use later
